API

PSIS.psis! — Method

psis!(args, r_eff; kwargs...)

In-place compute Pareto smoothed importance sampling (PSIS) log weights.

See psis for an out-of-place version and for description of arguments and return values.

PSIS.psis — Method

psis(log_ratios, r_eff; kwargs...) -> (log_weights, k)

Compute Pareto smoothed importance sampling (PSIS) log weights [VehtariSimpson2021].

See psis! for a version that smoothes the ratios in-place.

Arguments

log_ratios: an array of logarithms of importance ratios, with one of the following sizes:

(ndraws,): a vector of draws for a single parameter from a single chain

(nparams, ndraws): a matrix of draws for a multiple parameter from a single chain

(nparams, ndraws, nchains): an array of draws for multiple parameters from multiple chains, e.g. as might be generated with Markov chain Monte Carlo.

r_eff: the ratio of effective sample size of log_ratios and the actual sample size, used to account for autocorrelation, e.g. due to Markov chain Monte Carlo. If the ratios are known to be uncorrelated, then provide r_eff=ones(nparams).

Keywords

sorted=issorted(vec(log_ratios)): whether log_ratios are already sorted. Only accepted if nparams==1.

normalize=false: whether to normalize the log weights so that the resulting weights for a given parameter sum to one.

improved=false: If true, use the adaptive empirical prior of [Zhang2010]. If false, use the simpler prior of [ZhangStephens2009], which is also used in [VehtariSimpson2021].

Returns

log_weights: an array of smoothed log weights of the same size as log_ratios

k: for each parameter, the estimated shape parameter $k$ of the generalized Pareto distribution, which is useful for diagnosing the distribution of importance ratios. See details below.

Diagnostic

The shape parameter $k$ of the generalized Pareto distribution can be used to diagnose reliability and convergence of estimates using the importance weights [VehtariSimpson2021]:

if $k < \frac{1}{3}$, importance sampling is stable, and importance sampling (IS) and PSIS both are reliable.

if $k < \frac{1}{2}$, then the importance ratio distributon has finite variance, and the central limit theorem holds. As $k$ approaches the upper bound, IS becomes less reliable, while PSIS still works well but with a higher RMSE.

if $\frac{1}{2} ≤ k < 0.7$, then the variance is infinite, and IS can behave quite poorly. However, PSIS works well in this regime.

if $0.7 ≤ k < 1$, then it quickly becomes impractical to collect enough importance weights to reliably compute estimates, and importance sampling is not recommended.

if $k ≥ 1$, then neither the variance nor the mean of the raw importance ratios exists. The convergence rate is close to zero, and bias can be large with practical sample sizes.

A warning is raised if $k ≥ 0.7$.

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]

ZhangStephens2009Jin Zhang & Michael A. Stephens (2009) A New and Efficient Estimation Method for the Generalized Pareto Distribution, Technometrics, 51:3, 316-325, DOI: 10.1198/tech.2009.08017

Zhang2010Jin Zhang (2010) Improving on Estimation for the Generalized Pareto Distribution, Technometrics, 52:3, 335-339, DOI: 10.1198/TECH.2010.09206