Internal

PSIS.GeneralizedPareto — Type

GeneralizedPareto{T<:Real}

The (zero-centered) generalized Pareto distribution.

Constructor

GeneralizedPareto(σ, k)

Construct the generalized Pareto distribution (GPD) with scale parameter $σ$ and shape parameter $k$. Note that this $k$ is equal to the commonly used shape parameter $ξ$. This is the same parameterization used by [VehtariSimpson2021] and is related to that used by [ZhangStephens2009] as $k \mapsto -k$.

PSIS.fit — Method

fit(T::Type{<:GeneralizedPareto}, x; kwargs...) -> ::T

Estimate a generalized Pareto distribution (GPD) given the points x.

Compute an empirical Bayes estimate of the parameters of the (zero-centered) GPD given the data x using the method described in [ZhangStephens2009]. The method estimates $θ = \frac{σ}{k}$

Keywords

sorted=false: whether x is already sorted

min_points=30: default number of points to use to compute estimate. Instead of min_points=20 as recommended by [ZhangStephens2009], we use 30 points as in the loo software. [ZhangStephens2009] notes that the estimator is not sensitive to this choice.

adjust_prior=true: whether to apply the weakly informative Gaussian prior on k suggested by [VehtariSimpson2021] to reduce variance in the estimate of k.

PSIS.quantile — Method

quantile(d::GeneralizedPareto, p)

Compute the $p$-quantile of the generalized Pareto distribution d.

ZhangStephens2009Zhang J & Stephens M A. (2009). A new and efficient estimation method for the generalized Pareto distribution, Technometrics, 51:3, 316-325, DOI: 10.1198/tech.2009.08017

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]