API

PSIS.psis — Function

psis(log_ratios, r_eff; kwargs...) -> (log_weights, k)

Compute Pareto smoothed importance sampling (PSIS) log weights [VehtariSimpson2021].

See psis! for version that smoothes the ratios in-place.

Arguments

log_ratios: an array of logarithms of importance ratios.

r_eff: the ratio of effective sample size of log_ratios and the actual sample size, used to correct for autocorrelation due to MCMC. r_eff=1 should be used if the ratios were sampled independently.

Keywords

sorted=issorted(log_ratios): whether log_ratios are already sorted.

normalize=false: whether to normalize the log weights so that sum(exp.(low_weights)) ≈ 1.

Returns

log_weights: an array of smoothed log weights

k: the estimated shape parameter $k$ of the generalized Pareto distribution, which is useful for diagnosing the distribution of importance ratios. See details below.

Diagnostic

The shape parameter $k$ of the generalized Pareto distribution can be used to diagnose reliability and convergence of estimates using the importance weights [VehtariSimpson2021]:

if $k < \frac{1}{3}$, importance sampling is stable, and importance sampling (IS) and PSIS both are reliable.

if $k < \frac{1}{2}$, then the importance ratio distributon has finite variance, and the central limit theorem holds. As $k$ approaches the upper bound, IS becomes less reliable, while PSIS still works well but with a higher RMSE.

if $\frac{1}{2} ≤ k < 0.7$, then the variance is infinite, and IS can behave quite poorly. However, PSIS works well in this regime.

if $0.7 ≤ k < 1$, then it quickly becomes impractical to collect enough importance weights to reliably compute estimates, and importance sampling is not recommended.

if $k ≥ 1$, then neither the variance nor the mean of the raw importance ratios exists. The convergence rate is close to zero, and bias can be large with practical sample sizes.

A warning is raised if $k ≥ 0.7$.

PSIS.psis! — Function

psis!(args...; kwargs...)

In-place compute Pareto smoothed importance sampling (PSIS) log weights.

See psis for an out-of-place version and for description of arguments and return values.

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]