Internal

PSIS.GeneralizedPareto — Type

GeneralizedPareto{T<:Real}

The generalized Pareto distribution.

This is equivalent to Distributions.GeneralizedPareto and can be converted to one with convert(Distributions.GeneralizedPareto, d).

Constructor

GeneralizedPareto(μ, σ, k)

Construct the generalized Pareto distribution (GPD) with location parameter $μ$, scale parameter $σ$ and shape parameter $k$.

Note

The shape parameter $k$ is equivalent to the commonly used shape parameter $ξ$. This is the same parameterization used by [VehtariSimpson2021] and is related to that used by [ZhangStephens2009] as $k \mapsto -k$.

PSIS.fit_gpd — Method

fit_gpd(x; μ=0, kwargs...)

Fit a GeneralizedPareto with location μ to the data x.

The fit is performed using the Empirical Bayes method of [ZhangStephens2009].

Keywords

prior_adjusted::Bool=true, If true, a weakly informative Normal prior centered on $\frac{1}{2}$ is used for the shape $k$.

sorted::Bool=issorted(x): If true, x is assumed to be sorted. If false, a sorted copy of x is made.

min_points::Int=30: The minimum number of quadrature points to use when estimating the posterior mean of $\theta = \frac{\xi}{\sigma}$.

ZhangStephens2009Jin Zhang & Michael A. Stephens (2009) A New and Efficient Estimation Method for the Generalized Pareto Distribution, Technometrics, 51:3, 316-325, DOI: 10.1198/tech.2009.08017